refactor: organize audit UI and extension surfaces
