+++ features/audit/ui/degree-side-panel.tsx
