+++ features/planner/degree-planner-page.tsx
@@ -2,7 +2,7 @@ import { HStack, VStack } from "@/components/ui/stack";
 import Title from "@/components/ui/text";
 import "@/entrypoints/styles/content.css";
 import { CourseSearchPanel } from "@/features/course-search/course-search-panel";
-import DegreeSidePanel from "@/features/degree-audit-app/shared/degree-side-panel";
+import DegreeSidePanel from "@/features/audit/ui/degree-side-panel";
 import SemesterDropdowns from "./semester-dropdowns";
 
 const MainContent = () => {
